+++ packages/keyring-api/src/api/v2/keyring-type.ts
@@ -47,5 +47,5 @@ export enum KeyringType {
   /**
    * Represents keyring for cash accounts.
    */
-  CashAccount = 'cash-account',
+  Cash = 'cash',
 }