+++ packages/keyring-eth-cash-account/src/index.ts
@@ -1,1 +1,0 @@
-export { CashAccountKeyring } from './cash-account-keyring';