+++ packages/keyring-eth-cash/src/index.ts
@@ -1,0 +1,1 @@
+export { CashKeyring } from './cash-keyring';